feat: implement esm.js task
